Move all sinks to their own packages, refactor tests accordingly.
