+++ sinks/datadog/datadog_test.go
@@ -1,0 +1,137 @@
+package datadog
+
+import (
+	"net/http"
+	"testing"
+	"time"
+
+	"github.com/DataDog/datadog-go/statsd"
+	"github.com/sirupsen/logrus"
+	"github.com/stretchr/testify/assert"
+	"github.com/stripe/veneur/samplers"
+)
+
+// DDMetricsRequest represents the body of the POST request
+// for sending metrics data to Datadog
+// Eventually we'll want to define this symmetrically.
+type DDMetricsRequest struct {
+	Series []DDMetric
+}
+
+// assertMetrics checks that all expected metrics are present
+// and have the correct value
+func assertMetrics(t *testing.T, metrics DDMetricsRequest, expectedMetrics map[string]float64) {
+	// it doesn't count as accidentally quadratic if it's intentional
+	for metricName, expectedValue := range expectedMetrics {
+		assertMetric(t, metrics, metricName, expectedValue)
+	}
+}
+
+func assertMetric(t *testing.T, metrics DDMetricsRequest, metricName string, value float64) {
+	defer func() {
+		if r := recover(); r != nil {
+			assert.Fail(t, "error extracting metrics", r)
+		}
+	}()
+	for _, metric := range metrics.Series {
+		if metric.Name == metricName {
+			assert.Equal(t, int(value+.5), int(metric.Value[0][1]+.5), "Incorrect value for metric %s", metricName)
+			return
+		}
+	}
+	assert.Fail(t, "did not find expected metric", metricName)
+}
+
+func TestDatadogRate(t *testing.T) {
+	ddSink := DatadogMetricSink{
+		hostname: "somehostname",
+		tags:     []string{"a:b", "c:d"},
+		interval: 10,
+	}
+
+	metrics := []samplers.InterMetric{{
+		Name:      "foo.bar.baz",
+		Timestamp: time.Now().Unix(),
+		Value:     float64(10),
+		Tags:      []string{"gorch:frobble", "x:e"},
+		Type:      samplers.CounterMetric,
+	}}
+	ddMetrics := ddSink.finalizeMetrics(metrics)
+	assert.Equal(t, "rate", ddMetrics[0].MetricType, "Metric type should be rate")
+	assert.Equal(t, float64(1.0), ddMetrics[0].Value[0][1], "Metric rate wasnt computed correctly")
+}
+
+func TestServerTags(t *testing.T) {
+	ddSink := DatadogMetricSink{
+		hostname: "somehostname",
+		tags:     []string{"a:b", "c:d"},
+		interval: 10,
+	}
+
+	metrics := []samplers.InterMetric{{
+		Name:      "foo.bar.baz",
+		Timestamp: time.Now().Unix(),
+		Value:     float64(10),
+		Tags:      []string{"gorch:frobble", "x:e"},
+		Type:      samplers.CounterMetric,
+	}}
+
+	ddMetrics := ddSink.finalizeMetrics(metrics)
+	assert.Equal(t, "somehostname", ddMetrics[0].Hostname, "Metric hostname uses argument")
+	assert.Contains(t, ddMetrics[0].Tags, "a:b", "Tags should contain server tags")
+}
+
+func TestHostMagicTag(t *testing.T) {
+	ddSink := DatadogMetricSink{
+		hostname: "badhostname",
+		tags:     []string{"a:b", "c:d"},
+	}
+
+	metrics := []samplers.InterMetric{{
+		Name:      "foo.bar.baz",
+		Timestamp: time.Now().Unix(),
+		Value:     float64(10),
+		Tags:      []string{"gorch:frobble", "host:abc123", "x:e"},
+		Type:      samplers.CounterMetric,
+	}}
+
+	ddMetrics := ddSink.finalizeMetrics(metrics)
+	assert.Equal(t, "abc123", ddMetrics[0].Hostname, "Metric hostname should be from tag")
+	assert.NotContains(t, ddMetrics[0].Tags, "host:abc123", "Host tag should be removed")
+	assert.Contains(t, ddMetrics[0].Tags, "x:e", "Last tag is still around")
+}
+
+func TestDeviceMagicTag(t *testing.T) {
+	ddSink := DatadogMetricSink{
+		hostname: "badhostname",
+		tags:     []string{"a:b", "c:d"},
+	}
+
+	metrics := []samplers.InterMetric{{
+		Name:      "foo.bar.baz",
+		Timestamp: time.Now().Unix(),
+		Value:     float64(10),
+		Tags:      []string{"gorch:frobble", "device:abc123", "x:e"},
+		Type:      samplers.CounterMetric,
+	}}
+
+	ddMetrics := ddSink.finalizeMetrics(metrics)
+	assert.Equal(t, "abc123", ddMetrics[0].DeviceName, "Metric devicename should be from tag")
+	assert.NotContains(t, ddMetrics[0].Tags, "device:abc123", "Host tag should be removed")
+	assert.Contains(t, ddMetrics[0].Tags, "x:e", "Last tag is still around")
+}
+
+func TestNewDatadogSpanSinkConfig(t *testing.T) {
+	// test the variables that have been renamed
+	stats, _ := statsd.NewBuffered("localhost:1235", 1024)
+	ddSink, err := NewDatadogSpanSink("http://example.com", 100, stats, &http.Client{}, nil, logrus.New())
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = ddSink.Start(nil)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	assert.Equal(t, "http://trace", ddSink.traceAddress)
+}
